# src/main/resources/api/api.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080/openapi
    description: Generated server url
tags:
  - name: Project
    description: Project API
  - name: Issue
    description: Issue API
paths:
  /projects/{projectId}:
    get:
      tags:
        - Project
      operationId: getProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Project'
    put:
      tags:
        - Project
      operationId: updateProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      tags:
        - Project
      operationId: deleteProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    patch:
      tags:
        - Project
      operationId: patchProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Project'

  /projects:
    get:
      tags:
        - Project
      operationId: getProjects
      parameters:
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags:
        - Project
      operationId: createProject
      parameters:
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Project'

  /issues:
    get:
      tags:
        - Issue
      operationId: getIssues
      parameters:
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Issue'
    post:
      tags:
        - Issue
      operationId: createIssue
      parameters:
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Issue'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Issue'

  /issues/{issueId}:
    get:
      tags:
        - Issue
      operationId: getIssue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Issue'
    put:
      tags:
        - Issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Issue'
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Issue'

    delete:
      tags:
        - Issue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK


  /comments:
    get:
      tags:
        - Comment
      operationId: getComments
      parameters:
        - name: commentId
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Comment'
    post:
      tags:
        - Comment
      operationId: createComment
      parameters:
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Comment'

  /comments/{commentId}:
    get:
      tags:
        - Comment
      operationId: getComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Comment'
    put:
      tags:
        - Comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Comment'

    delete:
      tags:
        - Comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: X-User-Name
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK

components:
  schemas:
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int32
        description:
          type: string
        author:
          type: string
        issueId:
          type: integer
          format: int32
    Issue:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        author:
          type: string
        projectId:
          type: integer
          format: int32
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'

    Project:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        authorId:
          type: integer
          format: int32
        issues:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
    Person:
      type: object
      properties:
        id:
          type: integer
          format: int32
        username:
          type: string
        name:
          type: string
        surname:
          type: string